// frontend/src/state/actions/types.ts
enum ActionsTypes {
	MAININDICATORS_START = 'MAININDICATORS_START',
	MAININDICATORS_SUCCESS = 'MAININDICATORS_SUCCESS',
	MAININDICATORS_FAIL = 'MAININDICATORS_FAIL',

	PROFITANDLOST_START = 'PROFITANDLOST_START',
	PROFITANDLOST_SUCCESS = 'PROFITANDLOST_SUCCESS',
	PROFITANDLOST_FAIL = 'PROFITANDLOST_FAIL',

	CALCFOR_START = 'CALCFOR_START',
	CALCFOR_SUCCESS = 'CALCFOR_SUCCESS',
	CALCFOR_FAIL = 'CALCFOR_FAIL',

	GM_START = 'GM_START',
	GM_SUCCESS = 'GM_SUCCESS',
	GM_FAIL = 'GM_FAIL',

	CAPITAL_START = 'CAPITAL_START',
	CAPITAL_SUCCESS = 'CAPITAL_SUCCESS',
	CAPITAL_FAIL = 'CAPITAL_FAIL',

	INTERBANKDEPOSITS_START = 'INTERBANKDEPOSITS_START',
	INTERBANKDEPOSITS_SUCCESS = 'INTERBANKDEPOSITS_SUCCESS',
	INTERBANKDEPOSITS_FAIL = 'INTERBANKDEPOSITS_FAIL',

	LIQUIDITY_START = 'LIQUIDITY_START',
	LIQUIDITY_SUCCESS = 'LIQUIDITY_SUCCESS',
	LIQUIDITY_FAIL = 'LIQUIDITY_FAIL',

	PLAT_START = 'PLAT_START',
	PLAT_SUCCESS = 'PLAT_SUCCESS',
	PLAT_FAIL = 'PLAT_FAIL',

	TOPDEPOSITS_START = 'TOPDEPOSITS_START',
	TOPDEPOSITS_SUCCESS = 'TOPDEPOSITS_SUCCESS',
	TOPDEPOSITS_FAIL = 'TOPDEPOSITS_FAIL',

	TIMEDEPOSITS_START = 'TIMEDEPOSITS_START',
	TIMEDEPOSITS_SUCCESS = 'TIMEDEPOSITS_SUCCESS',
	TIMEDEPOSITS_FAIL = 'TIMEDEPOSITS_FAIL',

	TIMEDEPOCLIENTS_START = 'TIMEDEPOCLIENTS_START',
	TIMEDEPOCLIENTS_SUCCESS = 'TIMEDEPOCLIENTS_SUCCESS',
	TIMEDEPOCLIENTS_FAIL = 'TIMEDEPOCLIENTS_FAIL',

	DASHBOARD_START = 'DASHBOARD_START',
	DASHBOARD_SUCCESS = 'DASHBOARD_SUCCESS',
	DASHBOARD_FAIL = 'DASHBOARD_FAIL',

	CREDITDATA_START = 'CREDITDATA_START',
	CREDITDATA_SUCCESS = 'CREDITDATA_SUCCESS',
	CREDITDATA_FAIL = 'CREDITDATA_FAIL',

	DASHBOARD_ACTIVE_TAB_CHANGE = 'DASHBOARD_ACTIVE_TAB_CHANGE',

	CORRESPONDENT_START = 'CORRESPONDENT_START',
	CORRESPONDENT_SUCCESS = 'CORRESPONDENT_SUCCESS',

	CORRESPONDENT_FAIL = 'CORRESPONDENT_FAIL',
	CURRENCYPOSITION_START = 'CURRENCYPOSITION_START',
	CURRENCYPOSITION_SUCCESS = 'CURRENCYPOSITION_SUCCESS',
	CURRENCYPOSITION_FAIL = 'CURRENCYPOSITION_FAIL',

	OPERATIONAL_DAYS_START = 'OPERATIONAL_DAYS_START',
	OPERATIONAL_DAYS_SUCCESS = 'OPERATIONAL_DAYS_SUCCESS',
	OPERATIONAL_DAYS_FAIL = 'OPERATIONAL_DAYS_FAIL',

	LAST_UPDATE_START = 'LAST_UPDATE_START',
	LAST_UPDATE_SUCCESS = 'LAST_UPDATE_SUCCESS',
	LAST_UPDATE_FAIL = 'LAST_UPDATE_FAIL',

	LOGIN_START = 'LOGIN_START',
	LOGIN_SUCCESS = 'LOGIN_SUCCESS',
	LOGIN_FAIL = 'LOGIN_FAIL',

	ADDUSER_START = 'ADDUSER_START',
	ADDUSER_SUCCESS = 'ADDUSER_SUCCESS',
	ADDUSER_FAIL = 'ADDUSER_FAIL',

	GETUSER_START = 'GETUSER_START',
	GETUSER_SUCCESS = 'GETUSER_SUCCESS',
	GETUSER_FAIL = 'GETUSER_FAIL',

	EDITUSER_START = 'EDITUSER_START',
	EDITUSER_SUCCESS = 'EDITUSER_SUCCESS',
	EDITUSER_FAIL = 'EDITUSER_FAIL',
	EDITUSER_REFRESH = 'EDITUSER_REFRESH',

	ADDUSER_REFRESH = 'ADDUSER_REFRESH',

	USERS_START = 'USERS_START',
	USERS_SUCCESS = 'USERS_SUCCESS',
	USERS_FAIL = 'USERS_FAIL',

	LOGOUT = 'LOGOUT',

	CHANGE_DATE = 'CHANGE_DATE',
	DELETE_USER = 'DELETE_USER',
	UPDATE_CBN = 'UPDATE_CBN',

	USER_EXITED = 'USER_EXITED',

	CORRESPONDENT_CURRENT_UPDATE = 'CORRESPONDENT_CURRENT_UPDATE',
	LIQUIDITY_CURRENT_UPDATE = 'LIQUIDITY_CURRENT_UPDATE',

	CORRESPONDENT_CURRENT_START = 'CORRESPONDENT_CURRENT_START',
	CORRESPONDENT_CURRENT_SUCCESS = 'CORRESPONDENT_CURRENT_SUCCESS',
	CORRESPONDENT_CURRENT_FAIL = 'CORRESPONDENT_CURRENT_FAIL',

	LIQUIDITY_CURRENT_START = 'LIQUIDITY_CURRENT_START',
	LIQUIDITY_CURRENT_SUCCESS = 'LIQUIDITY_CURRENT_SUCCESS',
	LIQUIDITY_CURRENT_FAIL = 'LIQUIDITY_CURRENT_FAIL',

	DASHBOARDMONTHLY_START = 'DASHBOARDMONTHLY_START',
	DASHBOARDMONTHLY_SUCCESS = 'DASHBOARDMONTHLY_SUCCESS',
	DASHBOARDMONTHLY_FAIL = 'DASHBOARDMONTHLY_FAIL',

	DEPOSITSBYDEADLINE_START = 'DEPOSITSBYDEADLINE_START',
	DEPOSITSBYDEADLINE_SUCCESS = 'DEPOSITSBYDEADLINE_SUCCESS',
	DEPOSITSBYDEADLINE_FAIL = 'DEPOSITSBYDEADLINE_FAIL',

	REPORTLIABILITIES_START = 'REPORTLIABILITIES_START',
	REPORTLIABILITIES_SUCCESS = 'REPORTLIABILITIES_SUCCESS',
	REPORTLIABILITIES_FAIL = 'REPORTLIABILITIES_FAIL',

	GAP_START = 'GAP_START',
	GAP_SUCCESS = 'GAP_SUCCESS',
	GAP_FAIL = 'GAP_FAIL',

	GAP_LAST_UPDATE_START = 'GAP_LAST_UPDATE_START',
	GAP_LAST_UPDATE_SUCCESS = 'GAP_LAST_UPDATE_SUCCESS',
	GAP_LAST_UPDATE_FAIL = 'GAP_LAST_UPDATE_FAIL',

	BANKLIMITS_START = 'BANKLIMITS_START',
	BANKLIMITS_SUCCESS = 'BANKLIMITS_SUCCESS',
	BANKLIMITS_FAIL = 'BANKLIMITS_FAIL',

	GAPMANUAL_START = 'GAPMANUAL_START',
	GAPMANUAL_SUCCESS = 'GAPMANUAL_SUCCESS',
	GAPMANUAL_FAIL = 'GAPMANUAL_FAIL',

	FCRB_START = 'FCRB_START',
	FCRB_SUCCESS = 'FCRB_SUCCESS',
	FCRB_FAIL = 'FCRB_FAIL',

	NOSTROMATRIX_START = 'NOSTROMATRIX_START',
	NOSTROMATRIX_SUCCESS = 'NOSTROMATRIX_SUCCESS',
	NOSTROMATRIX_FAIL = 'NOSTROMATRIX_FAIL',

	VLABUFFER_START = 'VLABUFFER_START',
	VLABUFFER_SUCCESS = 'VLABUFFER_SUCCESS',
	VLABUFFER_FAIL = 'VLABUFFER_FAIL',

	FILIALEFFECTIVENESS_START = 'FILIALEFFECTIVENESS_START',
	FILIALEFFECTIVENESS_SUCCESS = 'FILIALEFFECTIVENESS_SUCCESS',
	FILIALEFFECTIVENESS_FAIL = 'FILIALEFFECTIVENESS_FAIL'
}